Move most of common module in swcc bin...
- backtrace mod and ColorChoice are now in common mod
- Add some first code checking stuff to repl execute_code
