+++ src/bin/common/mod.rs
@@ -1,183 +1,3 @@
-use ansi_term::{ANSIString, Colour};
-use git_testament::git_testament_macros;
-use pico_args::Arguments;
-use saltwater::{
-    data::{error::CompileWarning, Location},
-    Error, Files, Opt,
-};
-use std::{
-    collections::VecDeque,
-    ffi::OsStr,
-    num::NonZeroUsize,
-    path::PathBuf,
-    process,
-    sync::atomic::{AtomicUsize, Ordering},
-};
-
-static ERRORS: AtomicUsize = AtomicUsize::new(0);
-static WARNINGS: AtomicUsize = AtomicUsize::new(0);
-
-macro_rules! type_sizes {
-    ($($type: ty),* $(,)?) => {
-        $(println!("{}: {}", stringify!($type), std::mem::size_of::<$type>());)*
-    };
-}
-
-git_testament_macros!(version);
-
-pub(super) const HELP: &str = concat!(
-    env!("CARGO_PKG_NAME"), " ", env!("CARGO_PKG_VERSION"), "\n",
-    "Joshua Nelson <jyn514@gmail.com>\n",
-    env!("CARGO_PKG_DESCRIPTION"), "\n",
-    "Homepage: ", env!("CARGO_PKG_REPOSITORY"), "\n",
-    "\n",
-    "usage: ", env!("CARGO_PKG_NAME"), " [FLAGS] [OPTIONS] [<file>]
-
-FLAGS:
-        --debug-ast        If set, print the parsed abstract syntax tree (AST) in addition to compiling.
-                            The AST does no type checking or validation, it only parses.
-        --debug-hir        If set, print the high intermediate representation (HIR) in addition to compiling.
-                            This does type checking and validation and also desugars various expressions.
-        --debug-ir         If set, print the intermediate representation (IR) of the program in addition to compiling.
-        --debug-lex        If set, print all tokens found by the lexer in addition to compiling.
-        --jit              If set, will use JIT compilation for C code and instantly run compiled code (No files produced).
-                            NOTE: this option only works if saltwater was compiled with the `jit` feature.
-    -h, --help             Prints help information
-    -c, --no-link          If set, compile and assemble but do not link. Object file is machine-dependent.
-    -E, --preprocess-only  If set, preprocess only, but do not do anything else.
-                            Note that preprocessing discards whitespace and comments.
-                            There is not currently a way to disable this behavior.
-    -V, --version          Prints version information
-
-OPTIONS:
-        --color <when>       When to use color. May be \"never\", \"auto\", or \"always\". [default: auto]
-    -o, --output <output>    The output file to use. [default: a.out]
-        --max-errors <max>   The maximum number of errors to allow before giving up.
-                             Use 0 to allow unlimited errors. [default: 10]
-    -I, --include <dir>      Add a directory to the local include path (`#include \"file.h\"`).
-                              Can be specified multiple times to add multiple directories.
-    -D, --define <id[=val]>  Define an object-like macro.
-                              Can be specified multiple times to add multiple macros.
-                              `val` defaults to `1`.
-
-ARGS:
-    <file>    The file to read C source from. \"-\" means stdin (use ./- to read a file called '-').
-              Only one file at a time is currently accepted. [default: -]"
-);
-
-pub(super) const USAGE: &str = "\
-usage: swcc [--help | -h] [--version | -V] [--debug-ir] [--debug-ast] [--debug-lex]
-           [--debug-hir] [--jit] [--no-link | -c] [--preprocess-only | -E]
-           [-I <dir>] [-D <id[=val]>] [<file>]";
-
-pub(super) fn parse_args() -> Result<(BinOpt, PathBuf), pico_args::Error> {
-    use std::collections::HashMap;
-
-    let mut input = Arguments::from_env();
-    if input.contains("-h") {
-        println!("{}", USAGE);
-        std::process::exit(1);
-    } else if input.contains("--help") {
-        println!("{}", HELP);
-        std::process::exit(1);
-    }
-    if input.contains(["-V", "--version"]) {
-        println!("{} {}", env!("CARGO_PKG_NAME"), version_testament!());
-        std::process::exit(0);
-    }
-    if input.contains("--print-type-sizes") {
-        use saltwater::data::*;
-        type_sizes!(
-            Location,
-            CompileError,
-            Type,
-            ast::Expr,
-            ast::ExprType,
-            hir::Expr,
-            hir::ExprType,
-            ast::Stmt,
-            ast::StmtType,
-            hir::Stmt,
-            hir::StmtType,
-            ast::Declaration,
-            hir::Declaration,
-            hir::Variable,
-            StructType,
-            Token,
-        );
-    }
-    let output = input
-        .opt_value_from_os_str(["-o", "--output"], os_str_to_path_buf)?
-        .unwrap_or_else(|| "a.out".into());
-    let max_errors = input
-        .opt_value_from_fn("--max-errors", |s| {
-            usize::from_str_radix(s, 10).map(NonZeroUsize::new)
-        })?
-        .unwrap_or_else(|| Some(NonZeroUsize::new(10).unwrap()));
-    let color_choice = input
-        .opt_value_from_str("--color")?
-        .unwrap_or(ColorChoice::Auto);
-    let mut search_path = Vec::new();
-    while let Some(include) =
-        input.opt_value_from_os_str(["-I", "--include"], os_str_to_path_buf)?
-    {
-        search_path.push(include);
-    }
-    let mut definitions = HashMap::new();
-    while let Some(arg) = input.opt_value_from_str::<_, String>(["-D", "--define"])? {
-        use pico_args::Error::ArgumentParsingFailed;
-        use saltwater::data::error::LexError;
-        use std::convert::TryInto;
-
-        let mut iter = arg.splitn(2, '=');
-        let key = iter
-            .next()
-            .expect("apparently I don't understand pico_args");
-        let val = iter.next().unwrap_or("1");
-        let def = val
-            .try_into()
-            .map_err(|err: LexError| ArgumentParsingFailed {
-                cause: err.to_string(),
-            })?;
-        definitions.insert(key.into(), def);
-    }
-
-    let bin_opt = BinOpt {
-        preprocess_only: input.contains(["-E", "--preprocess-only"]),
-        opt: Opt {
-            debug_lex: input.contains("--debug-lex"),
-            debug_asm: input.contains("--debug-ir"),
-            debug_ast: input.contains("--debug-ast"),
-            debug_hir: input.contains("--debug-hir"),
-            no_link: input.contains(["-c", "--no-link"]),
-            #[cfg(feature = "jit")]
-            jit: input.contains("--jit"),
-            max_errors,
-            definitions,
-            search_path,
-            // This is a little odd because `free` expects no arguments to be left,
-            // so we have to parse it last.
-            filename: input
-                .free_from_os_str(os_str_to_path_buf)?
-                .unwrap_or_else(|| PathBuf::from("-")),
-        },
-        color: color_choice,
-    };
-    Ok((bin_opt, output))
-}
-
-pub(super) struct BinOpt {
-    /// The options that will be passed to `compile()`
-    pub(super) opt: Opt,
-    /// If set, preprocess only, but do not do anything else.
-    ///
-    /// Note that preprocessing discards whitespace and comments.
-    /// There is not currently a way to disable this behavior.
-    pub(super) preprocess_only: bool,
-    /// Whether or not to use color
-    pub(super) color: ColorChoice,
-}
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(super) enum ColorChoice {
     Always,
@@ -207,148 +27,6 @@ impl std::str::FromStr for ColorChoice {
     }
 }
 
-fn os_str_to_path_buf(os_str: &OsStr) -> Result<PathBuf, bool> {
-    Ok(os_str.into())
-}
-
-pub(super) fn err_exit(
-    err: Error,
-    max_errors: Option<NonZeroUsize>,
-    color: ColorChoice,
-    files: &Files,
-) -> ! {
-    use Error::*;
-    match err {
-        Source(errs) => {
-            for err in &errs {
-                error(&err.data, err.location(), files, color);
-            }
-            if let Some(max) = max_errors {
-                if usize::from(max) <= errs.len() {
-                    println!(
-                        "fatal: too many errors (--max-errors {}), stopping now",
-                        max
-                    );
-                }
-            }
-            let (num_warnings, num_errors) = (get_warnings(), get_errors());
-            print_issues(num_warnings, num_errors);
-            process::exit(2);
-        }
-        IO(err) => fatal(&err, 3, color),
-        Platform(err) => fatal(&err, 4, color),
-    }
-}
-
-pub(super) fn print_issues(warnings: usize, errors: usize) {
-    if warnings == 0 && errors == 0 {
-        return;
-    }
-    let warn_msg = if warnings > 1 { "warnings" } else { "warning" };
-    let err_msg = if errors > 1 { "errors" } else { "error" };
-    let msg = match (warnings, errors) {
-        (0, _) => format!("{} {}", errors, err_msg),
-        (_, 0) => format!("{} {}", warnings, warn_msg),
-        (_, _) => format!("{} {} and {} {}", warnings, warn_msg, errors, err_msg),
-    };
-    eprintln!("{} generated", msg);
-}
-
-pub(super) fn error<T: std::fmt::Display>(
-    msg: T,
-    location: Location,
-    file_db: &Files,
-    color: ColorChoice,
-) {
-    ERRORS.fetch_add(1, Ordering::Relaxed);
-    let prefix = if color.use_color_for(atty::Stream::Stdout) {
-        Colour::Red.bold().paint("error")
-    } else {
-        ANSIString::from("error")
-    };
-    print!("{}", pretty_print(prefix, msg, location, file_db,));
-}
-
-#[must_use]
-pub(super) fn pretty_print<T: std::fmt::Display>(
-    prefix: ANSIString,
-    msg: T,
-    location: Location,
-    file_db: &Files,
-) -> String {
-    let file = location.file;
-    let start = file_db
-        .location(file, location.span.start)
-        .expect("start location should be in bounds");
-    let buf = format!(
-        "{}:{}:{} {}: {}\n",
-        file_db.name(file).to_string_lossy(),
-        start.line.number(),
-        start.column.number(),
-        prefix,
-        msg
-    );
-    // avoid printing spurious newline for errors and EOF
-    if location.span.end == 0 {
-        return buf;
-    }
-    let end = file_db
-        .location(file, location.span.end)
-        .expect("end location should be in bounds");
-    if start.line == end.line {
-        let line = file_db
-            .line_span(file, start.line)
-            .expect("line should be in bounds");
-        format!(
-            "{}{}{}{}\n",
-            buf,
-            file_db.source_slice(file, line).unwrap(),
-            " ".repeat(start.column.0 as usize),
-            "^".repeat((end.column - start.column).0 as usize)
-        )
-    } else {
-        buf
-    }
-}
-
-pub(super) fn handle_warnings(
-    warnings: VecDeque<CompileWarning>,
-    file_db: &Files,
-    color: ColorChoice,
-) {
-    WARNINGS.fetch_add(warnings.len(), Ordering::Relaxed);
-    let tag = if color.use_color_for(atty::Stream::Stdout) {
-        Colour::Yellow.bold().paint("warning")
-    } else {
-        ANSIString::from("warning")
-    };
-    for warning in warnings {
-        print!(
-            "{}",
-            pretty_print(tag.clone(), warning.data, warning.location, file_db)
-        );
-    }
-}
-
-#[inline]
-pub(super) fn get_warnings() -> usize {
-    WARNINGS.load(Ordering::SeqCst)
-}
-
-#[inline]
-pub(super) fn get_errors() -> usize {
-    ERRORS.load(Ordering::SeqCst)
-}
-
-pub(super) fn fatal<T: std::fmt::Display>(msg: T, code: i32, color: ColorChoice) -> ! {
-    if color.use_color_for(atty::Stream::Stderr) {
-        eprintln!("{}: {}", Colour::Black.bold().paint("fatal"), msg);
-    } else {
-        eprintln!("fatal: {}", msg);
-    }
-    process::exit(code);
-}
-
 #[cfg(feature = "color-backtrace")]
 pub(super) mod backtrace {
     use super::ColorChoice;
@@ -367,32 +45,5 @@ pub(super) mod backtrace {
 
     pub(crate) fn install(color: ColorChoice) {
         BacktracePrinter::new().install(Box::new(StandardStream::stderr(color.into())));
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::{Files, Location};
-    use ansi_term::Style;
-    use saltwater::data::lex::Span;
-
-    fn pp<S: Into<Span>>(span: S, source: &str) -> String {
-        let mut file_db = Files::new();
-        let source = String::from(source).into();
-        let file = file_db.add("<test-suite>", source);
-        let location = Location {
-            file,
-            span: span.into(),
-        };
-        let ansi_str = Style::new().paint("");
-        super::pretty_print(ansi_str, "", location, &file_db)
-    }
-    #[test]
-    fn pretty_print() {
-        assert_eq!(
-            dbg!(pp(8..15, "int i = \"hello\";\n")).lines().nth(2),
-            Some("        ^^^^^^^")
-        );
-        pp(0..0, "");
     }
 }
